move init_console_logger to babylon/core/logger/
