+++ src/BabylonCpp/include/babylon/core/logging/init_console_logger.h
@@ -1,0 +1,9 @@
+#include <babylon/babylon_api.h>
+
+namespace BABYLON
+{  
+  /**
+  * @brief Initializes the logging on the console
+  */
+  void BABYLON_SHARED_EXPORT initConsoleLogger();
+}
