--- src/SamplesRunner/console_logger.h
@@ -1,7 +1,0 @@
-namespace BABYLON
-{
-  namespace impl
-  {
-    void initializeConsoleLogger();
-  }
-}
